chore: add helper files and modify install script
- Add a new file `gen-docker-compose.helper` with the content of the diff
- Add a new file `gen-makefile.helper` with the content of the diff
- Modify the file `install.sh` - add lines to `~/.gitconfig` and `~/.config/nvim` if they don't exist
- Modify the file `install.sh` - add lines to `~/.zprofile` if it exists

+++ install.sh
@@ -1,6 +1,19 @@
 #!/usr/bin/env bash
 
-ln -vsf ~/yuan-vim/gitconfig ~/.gitconfig
-ln -vsf ~/yuan-vim/nvim ~/.config/nvim
+if [ ! -e ~/.gitconfig ]; then
+  ln -vsf ~/yuan-vim/gitconfig ~/.gitconfig
+fi
+
+if [ ! -e ~/.config/nvim ]; then
+  ln -vsf ~/yuan-vim/nvim ~/.config/nvim
+fi
+
+if [ -f ~/.zprofile ]; then
+  for helper in ~/yuan-vim/helper/*.helper; do
+  cat << EOF >> ~/.zprofile
+source ${helper}
+EOF
+  done
+fi
 
 echo "Install Success !"
